+++ packages/icons/CHANGELOG.md
@@ -8,6 +8,16 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 5.19.0 (6/9/2026 PST)
+
+#### 🚀 Updates
+
+- Feat: Publish icons 2026-06-09. [[#747](https://github.com/coinbase/cds/pull/747)]
+
+##### ⭐️ Added (1)
+
+- learningCenter
+
 ## 5.18.0 (5/29/2026 PST)
 
 #### 🚀 Updates